Affichage dynamique des températures horaires & refonte UI
- Ajout de la gestion des prévisions horaires dans TemperatureBar via la propriété ForecastDatas (conversion automatique Kelvin → Celsius, gestion des valeurs nulles).
- Adaptation du type Values en double?[] pour supporter l’absence de données.
- Ajout des méthodes GetForecastsForDay et GetCountForecastForDay dans InfoClimatManager pour récupérer les prévisions d’un jour donné.
- Refonte des dégradés de couleurs (passage à des tons verts) et nettoyage/ajout de ressources dans Colors.xaml et Common.xaml.
- Correction de la normalisation dans ColorHelper.GetColorMultipoint.
- Mise à jour de l’affichage min/max (« à » au lieu de « - ») sur les cartes météo.
- Documentation XML ajoutée sur les nouvelles propriétés et méthodes.
- Intégration complète dans MainWindow pour afficher la courbe des températures horaires du jour courant, avec gestion des cas sans données.

+++ AppCommon/Helpers/ColorHelper.cs
@@ -17,7 +17,9 @@ namespace AppCommon.Helpers
         /// </returns>
         public static Color GetColorMultipoint(double temperature, double minTemp, double maxTemp)
         {
-            var normalized = Math.Clamp((temperature - minTemp) / (maxTemp - minTemp), 0.0, 1.0);
+            // Normalize the temperature within the specified range.
+            // The null check above ensures 'temperature' is not null here, so direct conversion is safe.
+            var normalized = Math.Clamp(((double)temperature - minTemp) / (maxTemp - minTemp), 0.0, 1.0);
 
             // Define key colors
             Color[] colors =
